Allow searches for ICD19 code stem
The Locator search GUI has a built in autocompletion, that allows you
to enter the stem part of an ICD10 code (E.g. "O10"). You can then
select a more exact code (e.g. "O10.9").

But what if you want all codes associated with the stem (e.g. "O10.1",
"O10.2", ... "O10.9")? To do this, you would need to build a query with
all of these codes, a weildly and time-consuming affair, which is
naturally also error-prone.

If a user deliberately sends an ICD10 code stem, the existing code will
return nothing, because it only works with exact matches.

Hence this patch. It has additional CQL that first checks for the case
that the user has sent an ICD10 code stem, such as "O10". If so, the
stem is checked against all codes, and any matches are further checked
to make sure that they are also ICD10.

If the search term does not fit the pattern for an ICD10 stem, it is
passed on to the original CQL, which should be more efficient, because
it does full-string matches rather than partial matches.